rust: new package
This new package provides the compiler for the Rust programming language.

Currently, only the host variant is built.

The internal build process is as follows:

 1. rustc-stage0, provided by rust-bootstrap, is used to build
    rustc-stage1.
 2. rust-stage1 builds the final Rust compiler (rust-stage2)
    and the standard library for the host architecture.
 3. the standard library for the target architecture is build.

The Rust compiler uses LLVM as its backend, compiled with support for
x86, ARM, PowerPC and MIPS architectures.

To properly build the standard library for the target, two files are
needed:

 - the target configuration: a Makefile fragment telling the build system
   about the cross-compiler and such.
 - the target specification: a JSON file describing the target
   architecture to the LLVM backend.

These files are generated with a Python script which takes info from the
Buildroot configuration: `rust-target-gen`.

When compiling Rust code with this compiler, the generated program only
depends on the target C library, as it is statically linked to the Rust
standard library.

Signed-off-by: Eric Le Bihan <eric.le.bihan.dev@free.fr>